backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "", email = "" }]
dependencies = [
  "celery[redis]",
  "cloudpathlib[s3]",
  "icecream",
  "fastapi[standard]",
  "langchain",
  "langchain-anthropic>=0.3.1",
  "langchain-community",
  "langchain-google-genai>=2.0.8",
  "langchain-huggingface",
  "langchain-ollama>=0.2.2",
  "langchain-openai",
  "langchain-postgres",
  "langchain-unstructured",
  "langgraph",
  "langgraph-checkpoint-postgres>=2.0.9",
  "onnx",
  "psycopg[binary,pool]",
  "pdfplumber>=0.11.4",
  "pydantic-settings>=2.7.0",
  "pyjwt",
  "python-dotenv",
  "redis",
  "unstructured[pdf]",
]
requires-python = "==3.12.*"
readme = "README.md"
license = { text = "Apache-2.0" }

[dependency-groups]
dev = ["pytest", "watchdog[watchmedo]"]
